finish_db

+++ app/Http/Controllers/Student/StudentController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers\Student ;
 
 use App\models\Student\Student;
+use App\models\Family\Family;
 use Illuminate\Http\Request;
+use App\models\Publics\From;
+use Illuminate\Support\Facades\DB;
 
 use App\Http\Controllers\Controller;
 
@@ -13,7 +16,7 @@ class StudentController extends Controller
     public function index()
     {
         $students = Student::all();
-
+        //dd($students);
         return view('Student.students.students',compact('students'));
     }
 
@@ -97,15 +100,91 @@ class StudentController extends Controller
 
     public function destroy(request $request)
     {
-               /* here we have sued the table students and searched about the id using the find and then delete the
+        /* here we have sued the table students and searched about the id using the find and then delete the
         id using the id note: we have passed the id from the show using the route */
+        $user = DB::table('students')->where('id', $request->id )->value('family_id');
+        //dd($user);
+        if ($user == null) {
         Student::find($request->id)->delete();
         /*after delete the student by id we will redirect the to show and we will path deleting msg ->with('DeleteMsg', 'You Have Deleted the Student Successfully')*/
         session()->flash('Delete', 'تم حذف القسم بنجاح ');
         return redirect(route('student.show'));
+        }
+        else {
+        $families = Family::find($user);
+        $x=0;
+        $x = $families->student_statu ;
+        -- $x ;
+        $families->student_statu = $x;
+        $families->save();
+        Student::find($request->id)->delete();
+        /*after delete the student by id we will redirect the to show and we will path deleting msg ->with('DeleteMsg', 'You Have Deleted the Student Successfully')*/
+        session()->flash('Delete', 'تم حذف العلاقة بين الطالب والعائلة بنجاح ');
+        return redirect(route('student.show'));    
+        }
+        
+
     }
 
-    public function register(){
-        return view('student.students.register');
+    public function register(){    
+        $check = From::all();
+
+        if($check->isEmpty())
+        {
+        $form = From::create([
+            'student_form' => 0,
+            'family_form' => 0,
+            'medical_form' => 0,
+     
+        ]);
+        $enable = From::find(1);
+        //dd($enable);
+
+        return view('student.students.register',compact('enable'));    
+        }
+        else {
+        $enable = From::find(1);
+        //dd($enable);
+        return view('student.students.register',compact('enable'));   
+
+        }
+
     }
+
+    public function enable(Request $request){
+       
+        $check = From::all();
+        if($check->isEmpty())
+        {
+        $form = From::create([
+            'student_form' => 0,
+            'family_form' => 0,
+            'medical_form' => 0,
+    
+        ]);
+        $this->validate($request,[
+            'enable' => 'required'
+         ]);
+         $enable = From::find('1');
+         $enable->student_form = $request->enable;
+
+         $enable->save();
+         session()->flash('Form', 'تم تعديل حالة قائمة تسجيل الطالب بنجاح ');
+         //redirect after adding and saving the data with success msg ->with('SuccessMsg', 'You Have added Student Successfully')
+         return redirect(route('student.show'));
+        }
+        else {
+        $this->validate($request,[
+            'enable' => 'required'
+         ]);
+         $enable = From::find('1');
+         $enable->student_form = $request->enable;
+
+         $enable->save();
+         session()->flash('Form', 'تم تعديل حالة قائمة تسجيل الطالب بنجاح ');
+         //redirect after adding and saving the data with success msg ->with('SuccessMsg', 'You Have added Student Successfully')
+         return redirect(route('student.show'));
+        }
+    }        
+
 }
